% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay-GEN.R
\name{deploy_static_site}
\alias{deploy_static_site}
\title{Deploy static website to local Git folder}
\usage{
deploy_static_site(
  from_path,
  to_path,
  clean_to_path = TRUE,
  commit_msg = "auto-deploy static website"
)
}
\arguments{
\item{from_path}{Path to the directory containing the static website files that are to be deployed. A character scalar.}

\item{to_path}{Path to the Git (sub)folder to which the static website files are to be deployed. A character scalar.}

\item{clean_to_path}{Whether or not to completely wipe \code{to_path} before deploying the new website files. Setting this to \code{TRUE} ensures there are no obsolete
files left over from previous deployments.}

\item{commit_msg}{The Git commit message used for the deployment. A character scalar.}
}
\value{
A vector of paths to the deployed files/folders, invisibly.
}
\description{
Deploy static website to local Git folder
}
