% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{gh_releases}
\alias{gh_releases}
\title{List releases from GitHub repository}
\usage{
gh_releases(
  owner,
  name,
  n = Inf,
  incl_pre_releases = FALSE,
  incl_drafts = FALSE
)
}
\arguments{
\item{owner}{GitHub repository owner (GitHub user or organisation). A character scalar.}

\item{name}{Repository name. A character scalar.}

\item{n}{Number of releases to return. A positive integer or \code{Inf}. If \code{Inf}, all releases are returned, otherwise only the most recent \code{n} releases. Note
that requesting many releases may result in multiple GitHub API calls, which can take a considerable amount of time.}

\item{incl_pre_releases}{Whether or not to include pre-releases in the result. Pre-releases are usually not deemed ready for production and may be unstable.
For more information, see the \href{https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository}{GitHub Docs}.}

\item{incl_drafts}{Whether or not to include draft releases in the result. For more information, see the \href{https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository}{GitHub Docs}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id}, \code{version_nr}, \code{is_pre_release}, and \code{is_draft}.
}
\description{
Uses \code{\link[gh:gh]{gh::gh()}} to fetch all available \href{https://docs.github.com/repositories/releasing-projects-on-github}{releases} of the specified GitHub repository
via \href{https://docs.github.com/en/rest/releases/releases#list-releases}{GitHub's RESTful API} and returns them as a \link[tibble:tbl_df-class]{tibble} containing the
columns \code{id}, \code{version_nr}, \code{is_pre_release} and \code{is_draft}.
}
\examples{
yay::gh_releases(owner = "jgm",
                 name = "pandoc",
                 n = 3L)
}
\seealso{
Other GitHub functions:
\code{\link{gh_dir_ls}()},
\code{\link{gh_release_latest}()},
\code{\link{gh_text_file}()},
\code{\link{gh_text_files}()}
}
\concept{gh}
