% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{gl_file_raw}
\alias{gl_file_raw}
\title{Read in a file from a GitLab repository}
\usage{
gl_file_raw(
  path,
  id_proj,
  ref = NULL,
  base_url = pal::pkg_config_val("gitlab_base_url"),
  token = pal::pkg_config_val("gitlab_token"),
  max_tries = 3L
)
}
\arguments{
\item{path}{File path, relative to the repository root. A character scalar.}

\item{id_proj}{GitLab project identifier. An integerish number.}

\item{ref}{\href{https://git-scm.com/docs/revisions#_specifying_revisions}{Git revision expression} matching the desired Git tree object, e.g. a ref name (branch,
tag, etc.), a commit identifier, or another symbolic reference like \code{"HEAD~10"}. A character scalar. Note that the GitLab API doesn't support every type of
revision expression.}

\item{base_url}{Base URL to the GitLab v4 RESTful API root.}

\item{token}{\href{https://docs.gitlab.com/ee/api/rest/authentication.html}{GitLab access token} used for authentication}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
A character scalar if \code{path} refers to a text file, otherwise a raw vector.
}
\description{
Downloads the file under the specified path from a GitLab repository via \href{https://docs.gitlab.com/ee/api/rest/}{GitLab's v4 RESTful API} and returns its
content.
}
\examples{
yay::gl_file_raw(path = "DESCRIPTION",
                 id_proj = 24995012) |>
  cat()
}
\seealso{
Other gitlabr: 
\code{\link{gl_file_write}()},
\code{\link{gl_url}()}
}
\concept{gitlabr}
