% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{gl_url}
\alias{gl_url}
\title{Assemble GitLab repository URL}
\usage{
gl_url(
  id_proj,
  ...,
  ref = gl_branch_default(id_proj = id_proj, base_url = base_url, token = token,
    max_tries = max_tries),
  force_branch_ref = FALSE,
  base_url = pal::pkg_config_val("gitlab_base_url"),
  token = pal::pkg_config_val("gitlab_token"),
  max_tries = 3L
)
}
\arguments{
\item{...}{Optional path components added to the repository base URL.}
}
\value{
A character scalar.
}
\description{
Assemble GitLab repository URL
}
\seealso{
Other gitlabr: 
\code{\link{gl_file_raw}()},
\code{\link{gl_file_write}()}
}
\concept{gitlabr}
\keyword{internal}
