% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{open_as_tmp_spreadsheet}
\alias{open_as_tmp_spreadsheet}
\title{Open as temporary spreadsheet}
\usage{
open_as_tmp_spreadsheet(x, format = c("csv", "xlsx"), ..., quiet = TRUE)
}
\arguments{
\item{x}{A dataframe or \link[tibble:tibble]{tibble}, or something coercible to.}

\item{format}{The spreadsheet format to use. One of
\itemize{
\item \code{"csv"} for a \href{https://en.wikipedia.org/wiki/Comma-separated_values}{comma-separated values} file written using \code{\link[readr:write_delim]{readr::write_csv()}}. The default.
\item \code{"xlsx"} for an \href{https://en.wikipedia.org/wiki/Office_Open_XML}{Office Open XML} file commonly used by Microsoft Excel 2007+, written using
\code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}.
}}

\item{...}{Further arguments passed on to \code{\link[readr:write_delim]{readr::write_csv()}} or \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}, depending on \code{format}.}

\item{quiet}{Whether or not to echo the command to open the temporary spreadsheet on the console before running it.}
}
\value{
\code{x} invisibly.
}
\description{
Writes an \R object – usually tabular data like a dataframe or \link[tibble:tibble]{tibble} – to a temporary spreadsheet and subsequently opens that
spreadsheet in the system's default application using \code{\link[xopen:xopen]{xopen::xopen()}}. The latter is usually equivalent to double-clicking on the file in a file browser.
}
\examples{
\dontrun{
yay::open_as_tmp_spreadsheet(mtcars,
                             format = "xlsx")}
}
\seealso{
Other data frame / tibble functions:
\code{\link{show_diff}()}
}
\concept{tibble}
