% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{porkbun_dns_records_delete}
\alias{porkbun_dns_records_delete}
\title{Delete Porkbun DNS records}
\usage{
porkbun_dns_records_delete(
  records,
  domain,
  api_key = funky::config_val("porkbun_api_key"),
  secret_api_key = funky::config_val("porkbun_secret_api_key"),
  max_tries = 3L
)
}
\arguments{
\item{records}{DNS records. A dataframe/tibble with the columns
\code{type}, \code{hostname}, \code{value}, \code{ttl}, \code{priority}, \code{weight}, \code{port}, \code{flag}, \code{tag} and \code{target}. The first three columns are mandatory,
columns not listed here are silently ignored.}

\item{domain}{Domain name to delete DNS records for. A character scalar.}

\item{api_key}{\href{https://porkbun.com/api/json/v3/documentation#Authentication}{Porkbun API key} used for authentication. A character scalar.}

\item{secret_api_key}{\href{https://porkbun.com/api/json/v3/documentation#Authentication}{Porkbun secret API key} used for authentication. A character scalar.}

\item{max_tries}{Maximum number of request attempts in case of an HTTP error. An integerish scalar. Retries are performed using exponential backoff and jitter, see \code{\link[httr2:req_retry]{httr2::req_retry()}} for details.}
}
\value{
A character vector of deleted record identifiers, invisibly.
}
\description{
Deletes DNS records on Porkbun for the specified \code{domain} using the
\href{https://porkbun.com/api/json/v3/documentation#DNS\%20Delete\%20Record\%20by\%20Domain\%20and\%20ID}{\verb{/api/json/v3/dns/delete/\{domain\}/\{dns_record_id\}}} endpoint
of Porkbun's API. DNS \code{records} must be provided as either a character vector of DNS record identifiers or a dataframe/tibble with an \code{id} column. Further
columns are silently ignored.
}
\examples{
\dontrun{
yay::porkbun_dns_records_delete(domain = "my.site",
                                records = "xyz123")

# The output of `netlify_dns_records_get()` can directly be fed. To delete all (!) non-NS records:
yay::porkbun_dns_records_get(domain = "my.site") |>
  dplyr::filter(type != "NS") |>
  yay::porkbun_dns_records_delete(domain = "my.site")}
}
\seealso{
Other Porkbun functions:
\code{\link{porkbun_dns_records_get}()},
\code{\link{porkbun_dns_records_set}()}
}
\concept{porkbun}
