% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{porkbun_dns_records_get}
\alias{porkbun_dns_records_get}
\title{Get Porkbun DNS records}
\usage{
porkbun_dns_records_get(
  domain,
  api_key = pal::pkg_config_val("porkbun_api_key"),
  secret_api_key = pal::pkg_config_val("porkbun_secret_api_key"),
  max_tries = 3L
)
}
\arguments{
\item{domain}{Domain name to retrieve DNS records for. A character scalar.}

\item{api_key}{Porkbun.com API key. A character scalar.}

\item{secret_api_key}{Porkbun.com secret API key. A character scalar.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id}, \code{type}, \code{hostname}, \code{value}, \code{ttl}, \code{priority}, \code{weight}, \code{port}, \code{flag}, \code{tag}, and \code{target}.
}
\description{
Retrieves DNS records from Porkbun for the specified \code{domain} using the
\href{https://porkbun.com/api/json/v3/documentation#DNS\%20Retrieve\%20Records\%20by\%20Domain\%20or\%20ID}{\verb{/api/json/v3/dns/retrieve/\{domain\}}} endpoint of Porkbun's
API.
}
\examples{
\dontrun{
yay::porkbun_dns_records_get(domain = "my.site")

# to write "portable" record keys to a TOML file `dns_records.toml` with a `records` table
# NOTE that the CLI tool `jsontoml` is required for this: https://github.com/pelletier/go-toml/
yay::porkbun_dns_records_get(domain = "my.site") |>
  # remove Porkbun-specific cols
  dplyr::select(any_of(yay:::cols_dns_records$key)) |>
  # convert to target list structure
  list(records = _) |>
  # convert to JSON
  jsonlite::toJSON(auto_unbox = TRUE,
                   pretty = TRUE) |>
  # convert JSON to TOML via external CLI
  system2(input = _,
          stdout = TRUE,
          command = "jsontoml") |>
  # remove leading newline
  _[-1L] |>
  # write TOML to file
  brio::write_lines(path = "dns_records.toml")}
}
\seealso{
Other Porkbun functions:
\code{\link{porkbun_dns_records_delete}()},
\code{\link{porkbun_dns_records_set}()}
}
\concept{porkbun}
