% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{show_diff}
\alias{show_diff}
\title{Determine the differences between two data frames/tibbles in tabular diff format}
\usage{
show_diff(
  x,
  y,
  ignore_order = FALSE,
  ignore_col_types = FALSE,
  ids = NULL,
  ask = TRUE,
  bypass_rstudio_viewer = FALSE,
  verbose = TRUE,
  max_diffs = 10L,
  diff_text = "{x_lbl} is different from {y_lbl}",
  ask_text = "Do you wish to display the changes in tabular diff format?",
  caption = "{x_lbl} → {y_lbl}",
  ...
)
}
\arguments{
\item{x}{The data frame / tibble to check for changes.}

\item{y}{The data frame / tibble that \code{x} should be checked against, i.e. the reference.}

\item{ignore_order}{Whether or not to ignore the order of columns and rows.}

\item{ignore_col_types}{Whether or not to distinguish similar column types. Currently, if set to \code{TRUE}, this will convert factors to characters and integers
to doubles before the comparison.}

\item{ids}{A character vector of column names that make up a \href{https://en.wikipedia.org/wiki/Primary_key}{primary key}, if known. If \code{NULL}, heuristics are
used to find a decent key (or a set of decent keys).}

\item{ask}{Whether or not to ask interactively if the resulting difference object should be opened in case \code{x} and \code{y} differ. If \code{FALSE}, it will be opened
right away. Only relevant if run \link[base:interactive]{interactively}.}

\item{bypass_rstudio_viewer}{If \code{TRUE}, \code{x} and \code{y} actually differ, and \code{ask} is set to \code{TRUE}, the resulting difference object will be
opened in the system's default web browser instead of RStudio's built-in viewer. Only relevant if run within RStudio.}

\item{verbose}{Whether or not to also output the differences detected by \code{\link[pal:is_equal_df]{pal::is_equal_df()}} to the console.}

\item{max_diffs}{The maximum number of differences shown on the console. Only relevant if \code{verbose = TRUE}.}

\item{diff_text}{The text to display on the console in case \code{x} and \code{y} differ. It is passed to \code{\link[glue:glue]{glue::glue()}} allowing its string interpolation syntax to be
used. A character scalar.}

\item{ask_text}{The text that is displayed when \code{ask = TRUE}. Ignored if \code{ask = FALSE}. A character scalar.}

\item{caption}{The caption of the rendered difference object. It is passed to \code{\link[glue:glue]{glue::glue()}} allowing its string interpolation syntax to be used. A character
scalar.}

\item{...}{Further arguments passed on to \code{\link[daff:diff_data]{daff::diff_data()}}, excluding \code{data}, \code{data_ref}, \code{ids}, \code{ordered}, and \code{columns_to_ignore}.}
}
\value{
A \link[daff:daff]{difference object}, invisibly. It could be rendered later using \code{\link[daff:render_diff]{daff::render_diff()}}, for example.
}
\description{
Compares two \link[base:data.frame]{data frames}/\link[tibble:tbl_df-class]{tibbles} (or two objects coercible to tibbles like
\link[base:matrix]{matrices}) and offers to inspect any differences in \href{https://paulfitz.github.io/daff-doc/spec.html}{tabular diff format} as neatly rendered
HTML.
}
\details{
This function is basically a convenience wrapper combining \code{\link[pal:is_equal_df]{pal::is_equal_df()}}, \code{\link[daff:diff_data]{daff::diff_data()}} and \code{\link[daff:render_diff]{daff::render_diff()}}. If run non-interactively or
\code{ask = FALSE}, the differences will be shown right away, otherwise the user will be asked on the console.

Note that in tabular diff format, only changes in the \emph{column content} of \code{x} and \code{y} are visible, meaning that the following properties and changes
therein won't be displayed:
\itemize{
\item column types (e.g. integer vs. double)
\item row names and other attributes
}
}
\examples{
\dontrun{
library(magrittr)

mtcars \%>\%
  dplyr::mutate(dplyr::across(c(cyl, gear),
                              ~ dplyr::if_else(. > 4, . * 2, .))) \%>\%
  yay::show_diff(mtcars)}
}
\seealso{
Other data frame / tibble functions:
\code{\link{open_as_tmp_spreadsheet}()}
}
\concept{tibble}
