% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{zotero_write_bib}
\alias{zotero_write_bib}
\title{Write Zotero library to bibliography file}
\usage{
zotero_write_bib(
  lib_id,
  path,
  collection_key = NULL,
  incl_children = FALSE,
  format = "csljson",
  force = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{lib_id}{Zotero user or group library identifier. An object as returned by \code{\link[zoterro:zotero_credentials]{zoterro::zotero_user_id()}} or \code{\link[zoterro:zotero_credentials]{zoterro::zotero_group_id()}}.}

\item{path}{Path to write the bibliography file to. A character scalar.}

\item{collection_key}{\href{https://www.zotero.org/support/dev/web_api/v3/basics#collections}{Collection key},
see \code{\link[zoterro:collections]{collections()}}. If \code{NULL}, all items in the library are returned.}

\item{incl_children}{Whether or not to include child items from
sub-collections beneath the \code{collection_key}'s or the library's (if
\code{collection_key = NULL}) top level.}

\item{format}{Bibliographic data format used for export. One of:
\itemize{
\item \code{"bibtex"},
\item \code{"biblatex"},
\item \code{"csljson"},
\item \code{"mods"},
\item \code{"refer"},
\item \code{"rdf_bibliontology"},
\item \code{"rdf_dc"},
\item \code{"rdf_zotero"},
\item \code{"ris"},
\item \code{"wikipedia"}
}

For details, see the \href{https://www.zotero.org/support/dev/web_api/v3/basics#export_formats}{relevant Zotero documentation}.}

\item{force}{Whether or not to enforce overwriting the bibliography file regardless of actual changes to the Zotero library since the last export.}

\item{show_progress}{Whether or not to output progress indication.}
}
\value{
\code{path}, invisibly.
}
\description{
Writes a file of the chosen bibliography \code{format} to \code{path} with items belonging to a Zotero user or group library, optionally limited to a specific
collection. Also writes a text file with suffix \code{.version} alongside \code{path} which stores the \href{https://www.zotero.org/support/dev/web_api/v3/syncing}{Zotero library version number} and allows for efficient caching.
}
\details{
If the Zotero library's content hasn't changed since the last export, nothing will be written to \code{path} and the performed API request will be significantly
faster. See section \emph{Caching} in \code{\link[zoterro:zotero_api]{zoterro::zotero_api()}} for further details.
}
\concept{zotero}
