% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{netlify_dns_records_delete}
\alias{netlify_dns_records_delete}
\title{Delete Netlify DNS records}
\usage{
netlify_dns_records_delete(records, domain, token, max_tries = 3L)
}
\arguments{
\item{records}{DNS records to delete. A character vector of record identifiers or a dataframe/tibble with an \code{id} column. Further columns are silently
ignored.}

\item{domain}{Domain name to delete DNS records for. This is translated into the corresponding Netlify DNS Zone. A character scalar.}

\item{token}{Personal access token used for authentication.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
A character vector of deleted record identifiers, invisibly.
}
\description{
Deletes DNS records on Netlify for the specified \code{domain} using the
\href{https://open-api.netlify.com/#tag/dnsZone/operation/deleteDnsRecord}{\code{deleteDnsRecord}} endpoint of \href{https://docs.netlify.com/api/get-started/}{Netlify's REST API}. DNS \code{records} must be provided as either a character vector of record identifiers or a dataframe/tibble with
an \code{id} column. Further columns are silently ignored.
}
\details{
Supported are the DNS record types \code{ALIAS}, \code{CAA}, \code{CNAME}, \code{MX}, \code{NS}, \code{SPF}, \code{SRV} and \code{TXT}. \code{A} and
\code{AAAA} (or \code{NETLIFY} and \code{NETLIFY6}, respectively) cannot be deleted via the API, but must be configured via \href{https://app.netlify.com/}{Netlify's web interface} instead.
}
\examples{
\dontrun{
yay::netlify_dns_records_delete(domain = "my.site",
                                records = "xyz123",
                                token = Sys.getenv("NETLIFY_PAT"))

# The output of `netlify_dns_records_get()` can be directly feed. To delete all (!) records:
yay::netlify_dns_records_get(domain = "my.site",
                             token = Sys.getenv("NETLIFY_PAT")) |>
  dplyr::filter(!managed) |>
  yay::netlify_dns_records_delete(domain = "my.site",
                                  token = Sys.getenv("NETLIFY_PAT"))}
}
\seealso{
Other Netlify functions:
\code{\link{netlify_dns_records_get}()},
\code{\link{netlify_dns_records_set}()}
}
\concept{netlify}
