% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{netlify_dns_records_get}
\alias{netlify_dns_records_get}
\title{Get Netlify DNS records}
\usage{
netlify_dns_records_get(domain, token, max_tries = 3L)
}
\arguments{
\item{domain}{Domain name to retrieve DNS records for. This is translated into the corresponding Netlify DNS Zone. A character scalar.}

\item{token}{Personal access token used for authentication.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{id}, \code{dns_zone_id}, \code{site_id}, \code{managed}, \code{type}, \code{hostname}, \code{value}, \code{ttl}, \code{priority}, \code{weight}, \code{port}, \code{flag}, and \code{tag}.
}
\description{
Retrieves DNS records from Netlify for the specified \code{domain} using the
\href{https://open-api.netlify.com/#tag/dnsZone/operation/getDnsRecords}{\code{getDnsRecords}} endpoint of \href{https://docs.netlify.com/api/get-started/}{Netlify's REST API}.
}
\examples{
\dontrun{
yay::netlify_dns_records_get(domain = "my.site",
                             token = Sys.getenv("NETLIFY_PAT"))

# to write the "settable" record keys to a TOML file `dns_records.toml` with a `records` table
# NOTE that the CLI tool `jsontoml` is required for this: https://github.com/pelletier/go-toml/
yay::netlify_dns_records_get(domain = "my.site",
                             token = Sys.getenv("NETLIFY_PAT")) |>
  dplyr::select(
    all_of(yay:::netlify_dns_record_cols$key[yay:::netlify_dns_record_cols$settable])
  ) |>
  list(records = _) |>
  jsonlite::toJSON(auto_unbox = TRUE,
                   pretty = TRUE) |>
  system2(input = _,
          stdout = TRUE,
          command = "jsontoml") |>
  _[-1L] |>
  brio::write_lines(path = "dns_records.toml")}
}
\seealso{
Other Netlify functions:
\code{\link{netlify_dns_records_delete}()},
\code{\link{netlify_dns_records_set}()}
}
\concept{netlify}
