% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{netlify_dns_records_set}
\alias{netlify_dns_records_set}
\title{Set Netlify DNS records}
\usage{
netlify_dns_records_set(domain, records, token, max_tries = 3L)
}
\arguments{
\item{domain}{The domain name to set DNS records for. This is translated into the corresponding Netlify DNS Zone. A character scalar.}

\item{records}{A dataframe/tibble with the columns
\code{"type"}, \code{"hostname"}, \code{"value"}, \code{"ttl"}, \code{"priority"}, \code{"weight"}, \code{"port"}, \code{"flag"} and \code{"tag"}. Further columns are silently ignored.}

\item{token}{Personal access token used for authentication.}

\item{max_tries}{Maximum number of attempts to retry in case of an HTTP error. An integerish scalar.}
}
\value{
A list of response objects, invisibly.
}
\description{
Sets DNS records on Netlify for the specified \code{domain} using the
\href{https://open-api.netlify.com/#tag/dnsZone/operation/createDnsRecord}{\code{createDnsRecord}} endpoint of \href{https://docs.netlify.com/api/get-started/}{Netlify's REST API}. DNS \code{records} must be provided as a dataframe/tibble with the columns
\code{"type"}, \code{"hostname"}, \code{"value"}, \code{"ttl"}, \code{"priority"}, \code{"weight"}, \code{"port"}, \code{"flag"} and \code{"tag"}. Further columns are silently ignored.
}
\details{
Supported are the DNS record types \code{netlify_dns_record_types |> pal::wrap_chr("\\x60") |> cli::ansi_collapse( last = " and ")}. \code{A} and
\code{AAAA} (or \code{NETLIFY} and \code{NETLIFY6}, respectively) cannot be set via the API and must be configured via \href{https://app.netlify.com/}{Netlify's web interface}.
}
\examples{
\dontrun{
tibble::tribble(
  ~type,   ~hostname,            ~value,           ~ttl, ~priority, ~weight, ~port, ~flag, ~tag,
  "CAA",   "my.site",            "letsencrypt.org", 3600L, NA,      NA,      NA,    NA,    NA,
  "CNAME", "autoconfig",         "mailbox.org",     3600L, NA,      NA,      NA,    NA,    NA,
  "MX",    "my.site",            "mxext1.mailbox.org", 3600L, NA,   NA,      NA,    NA,    NA,
  "SRV",   "_hkps._tcp.my.site", "pgp.mailbox.org",    3600L, 1L,   1L,      443L,  NA,    NA,
  "TXT",   "_mta-sts",           "v=STSv1; id=001",    3600L, NA,   NA,      NA,    NA,    NA) |>
  yay::netlify_dns_records_set(domain = "my.site",
                               token  = "XYZ1234")}
}
\concept{netlify}
