% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay.gen.R
\name{str_normalize}
\alias{str_normalize}
\title{Apply regular-expression-based text normalization to strings}
\usage{
str_normalize(
  string,
  rules = yay::regex_text_normalization,
  n_context_chrs = 20L,
  verbose = TRUE
)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{rules}{A \link[tibble:tibble]{tibble} of regular expression patterns and replacements. It must have the columns \code{pattern} and \code{replacement}. \code{pattern}
can optionally be a list column condensing multiple patterns to the same replacement rule. Patterns are interpreted as regular expressions as described
in \code{\link[stringi:about_search_regex]{stringi::stringi-search-regex()}}. Replacements are interpreted as-is, except that references of the form \verb{\\1}, \verb{\\2}, etc. will be replaced with the
contents of the respective matched group (created in patterns using \verb{()}). Pattern-replacement pairs are processed in the order given, meaning that first
listed pairs are applied before later listed ones.}

\item{n_context_chrs}{The (maximum) number of characters displayed around the actual \code{string} and its replacement. The number refers to a single side of
\code{string}/replacement, so the total number of context characters is at the maximum \code{2 * n_context_chrs}. Only relevant if \code{verbose = TRUE}.}

\item{verbose}{Whether or not to display replacements on the console.}
}
\value{
\code{path} invisibly.
}
\description{
Applies a set of regular-expression-based text normalization rules to one or more strings. All performed replacements are displayed on the console by default
(\code{verbose = TRUE}).
}
\examples{
"This kind of “text normalization” is e.g. useful to apply before feeding stuff to ‘Pandoc’" |>
  yay::str_normalize()
}
\seealso{
Regular expression rules: \code{\link{regex_text_normalization}} \code{\link{regex_file_normalization}}

Other string functions:
\code{\link{str_normalize_file}()},
\code{\link{str_replace_file}()},
\code{\link{str_replace_verbose}()}
}
\concept{string}
