% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yay-package.R
\docType{package}
\name{yay-package}
\alias{yay}
\alias{yay-package}
\title{yay: Delightful Convenience Functions}
\description{
A delightful mix of convenience functions designed to put a
    smile on your face during typical R workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/salim_b/r/pkgs/yay}
  \item Report bugs at \url{https://gitlab.com/salim_b/r/pkgs/yay/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
